requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
aiofiles==23.1.0
jinja2==3.1.2
python-dotenv==1.0.0
openai==1.8.0
